Factor out github fetching
